refactor: rename folder name from constant to constants
