+++ src/constants/devices.ts
