+++ src/constants/messageType.ts
